feat: adding to run primary and secondary instances

+++ build/webpack/plugins/block.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 const editorKeyRegex = /^editor_blocks_(?<srcElement>.+)$/i;
 
 class BlocksPlugin {
-  constructor({ directory, routes, outputPath }) {
+  constructor({ directory, routes, outputPath, emptyDirectoryOnStart = false }) {
     this.routes = routes;
     this.directory = directory;
     this.outputPath = outputPath;
@@ -11,7 +11,7 @@ class BlocksPlugin {
     this.filesToEmit = {};
 
     const destDir = `${this.directory}/blocks`;
-    if (fs.existsSync(destDir)) {
+    if (emptyDirectoryOnStart && fs.existsSync(destDir)) {
       fs.rmSync(destDir, { recursive: true });
     }
   }
